Move openkapsarc -> remote.openkapsarc; adjust CLI & tests
