+++ item/historical/cli.py
@@ -4,7 +4,6 @@ from tempfile import TemporaryDirectory
 import click
 from click import Group
 
-from item.openkapsarc import OpenKAPSARC
 from . import (
     SCRIPTS,
     fetch_source,
@@ -22,24 +21,6 @@ def fetch(source):
     """Retrieve raw data for SOURCE."""
     path = fetch_source(source)
     print(f'Retrieved {path}')
-
-
-@historical.command()
-@click.argument('server', default=None, required=False)
-def demo(server):
-    """Access the KAPSARC APIs at the given SERVER."""
-    ok = OpenKAPSARC(server)
-
-    print('List of all datasets:')
-    for ds in ok.datasets():
-        print(ds)
-
-    print("\n\nDatasets with the 'iTEM' keyword:")
-    for ds in ok.datasets(kw='item'):
-        print(ds)
-
-    print('\n\nData from one table in a branch in a repository:')
-    print(ok.table('modal-split-of-freight-transport'))
 
 
 @historical.command()
